# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar lectura y escritura de archivos multimedia (fotos) usando MinIO como almacenamiento de objetos.",
        "title": "Photo Bucket Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/folders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista las carpetas del usuario autenticado. Sin parentId, lista las carpetas raiz",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Listar carpetas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Listar el contenido de esta carpeta",
                        "name": "parentId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/folder.Folder"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "parentId no existe o no pertenece al usuario",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea una carpeta para el usuario autenticado, opcionalmente dentro de otra carpeta (parentId)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Crear una carpeta",
                "parameters": [
                    {
                        "description": "Datos de la carpeta",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/folder.CreateFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/folder.Folder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "parentId no existe o no pertenece al usuario",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "ya existe una carpeta con ese nombre en el mismo padre",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/folders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Obtener una carpeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carpeta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/folder.Folder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina la carpeta si esta vacia. Si tiene subcarpetas o archivos, devuelve 409",
                "tags": [
                    "folders"
                ],
                "summary": "Eliminar una carpeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carpeta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "sin contenido"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "la carpeta tiene subcarpetas o archivos",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Renombrar una carpeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carpeta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo nombre",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/folder.RenameFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/folder.Folder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "ya existe una carpeta con ese nombre en el mismo padre",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/folders/{id}/move": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cambia el padre de la carpeta. parentId vacio la mueve a la raiz",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "folders"
                ],
                "summary": "Mover una carpeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carpeta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo padre",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/folder.MoveFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/folder.Folder"
                        }
                    },
                    "400": {
                        "description": "parentId invalido o crearia un ciclo",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "ya existe una carpeta con ese nombre en el destino",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/apperror.ErrorCode"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "apperror.ErrorCode": {
            "type": "string",
            "enum": [
                "BAD_REQUEST",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "NOT_FOUND",
                "CONFLICT",
                "VALIDATION_ERROR",
                "INTERNAL_ERROR",
                "NO_METHOD",
                "TOO_MANY_REQUEST"
            ],
            "x-enum-varnames": [
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeConflict",
                "CodeValidation",
                "CodeInternal",
                "CodeNoMethod",
                "CodeTooManyRequest"
            ]
        },
        "apperror.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/apperror.ErrorBody"
                }
            }
        },
...
        "folder.CreateFolderRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "parentId": {
                    "type": "string"
                }
            }
        },
        "folder.Folder": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "folder.MoveFolderRequest": {
            "type": "object",
            "properties": {
                "parentId": {
                    "type": "string"
                }
            }
        },
        "folder.RenameFolderRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribir \"Bearer\" seguido de un espacio y el token JWT obtenido en /user/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}